pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mantisai"
version = "0.1.1"
description = "A Python package for transcribing audio files, summarizing text, and extracting information using Gemini AI and Pydantic."
requires-python = ">=3.9"
authors = [
    { name = "Paul Elliot", email = "paul@paulelliot.com" }
]
readme = "README.md"
license = { text = "Apache 2.0" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pydantic>=1.8",
    "google-generativeai",
    "requests",
    "python-dotenv",
    "yt_dlp",
    "tenacity>=8.0.0",
    "grpcio==1.60.1",  # Pin specific version to avoid warnings
]

[project.optional-dependencies]
dev = [
    "flake8",
    "black",
    "sphinx",
    "sphinx_rtd_theme",
    "sphinx-autodoc-typehints",
    "SpeechRecognition",
]

[project.scripts]
mantis = "mantis.cli:main"

[tool.hatch.build]
# Hatch settings for building your packages (if needed)

[tool.hatch.build.targets.wheel]
packages = ["mantis"]  # Tell Hatch to include the mantis directory

[tool.hatch.version]
path = "mantis/__init__.py"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
markers = [
    "local_only: marks tests that should only run locally (not in CI)",
]
